requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain
bs4
tiktoken
streamlit
google-search-results
wikipedia
playwright
watchdog
lxml